pyproject.toml
[project]
name = "omega-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=3.2.0",
    "pillow>=11.1.0",
    "sentencepiece>=0.2.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.5",
]
